moved code into own router mod
